Split invalid tests into individual test file
This ensures that if any test fails it will be reported as such. There's no such need for valid test as all tests should pass anyway.
