+++ src/test/resources/regression/verification/overflow/invalid/Overflow10.scala
@@ -1,0 +1,9 @@
+import leon.lang._
+
+object Overflow10 {
+
+  def foo10(x: Int, y: Int): Int = {
+    x*y
+  }
+
+}
