+++ src/test/resources/regression/verification/overflow/invalid/Overflow11.scala
@@ -1,0 +1,11 @@
+import leon.lang._
+
+object Overflow11 {
+
+  def foo11(): Int = {
+    val x = 500000
+    val y = 500000
+    x*y
+  }
+
+}
