+++ src/test/resources/regression/verification/overflow/invalid/Overflow12.scala
@@ -1,0 +1,9 @@
+import leon.lang._
+
+object Overflow12 {
+
+  def foo12(x: Int): Int = {
+    -x
+  }
+
+}
