+++ src/test/resources/regression/verification/overflow/invalid/Overflow13.scala
@@ -1,0 +1,10 @@
+import leon.lang._
+
+object Overflow13 {
+
+  def foo13(x: Int, y: Int) = {
+    require(y != 0)
+    x / y
+  }
+
+}
