+++ src/test/resources/regression/verification/overflow/invalid/Overflow2.scala
@@ -1,0 +1,9 @@
+import leon.lang._
+
+object Overflow2 {
+
+  def foo2(x: Int): Int = {
+    x + (-1)
+  }
+
+}
