+++ src/test/resources/regression/verification/overflow/invalid/Overflow3.scala
@@ -1,0 +1,9 @@
+import leon.lang._
+
+object Overflow3 {
+
+  def foo3(x: Int): Int = {
+    x - 1
+  }
+
+}
