+++ src/test/resources/regression/verification/overflow/invalid/Overflow4.scala
@@ -1,0 +1,9 @@
+import leon.lang._
+
+object Overflow4 {
+
+  def foo4(x: Int): Int = {
+    x - (-1)
+  }
+
+}
