+++ src/test/resources/regression/verification/overflow/invalid/Overflow5.scala
@@ -1,0 +1,9 @@
+import leon.lang._
+
+object Overflow5 {
+
+  def foo5(x: Int, y: Int): Int = {
+    x + y
+  }
+
+}
