+++ src/test/resources/regression/verification/overflow/invalid/Overflow6.scala
@@ -1,0 +1,9 @@
+import leon.lang._
+
+object Overflow6 {
+
+  def foo6(x: Int, y: Int): Int = {
+    x - y
+  }
+
+}
