+++ src/test/resources/regression/verification/overflow/invalid/Overflow7.scala
@@ -1,0 +1,9 @@
+import leon.lang._
+
+object Overflow7 {
+
+  def foo7(x: Int): Int = {
+    x*2
+  }
+
+}
