+++ src/test/resources/regression/verification/overflow/invalid/Overflow8.scala
@@ -1,0 +1,9 @@
+import leon.lang._
+
+object Overflow8 {
+
+  def foo8(x: Int): Int = {
+    x*3
+  }
+
+}
