+++ src/test/resources/regression/verification/overflow/invalid/Overflow9.scala
@@ -1,0 +1,9 @@
+import leon.lang._
+
+object Overflow9 {
+
+  def foo9(x: Int): Int = {
+    x*4
+  }
+
+}
