+++ src/test/resources/regression/verification/strictarithmetic/invalid/StrictArithmetic1.scala
@@ -2,20 +2,8 @@ import leon.lang._
 
 object StrictArithmetic1 {
 
-  def foo(x: Int, y: Int) = {
+  def foo1(x: Int, y: Int) = {
     x << y
-  }
-
-  def foo2(x: Int, y: Int) = {
-    x >> y
-  }
-
-  def foo3(x: Int, y: Int) = {
-    x >>> y
-  }
-
-  def foo4(x: Int, y: Int) = {
-    x % y
   }
 
 }
