+++ src/test/resources/regression/verification/strictarithmetic/invalid/StrictArithmetic2.scala
@@ -1,0 +1,10 @@
+import leon.lang._
+
+object StrictArithmetic2 {
+
+  def foo2(x: Int, y: Int) = {
+    x >> y
+  }
+
+}
+
