+++ src/test/resources/regression/verification/strictarithmetic/invalid/StrictArithmetic3.scala
@@ -1,0 +1,10 @@
+import leon.lang._
+
+object StrictArithmetic3 {
+
+  def foo3(x: Int, y: Int) = {
+    x >>> y
+  }
+
+}
+
