+++ src/test/resources/regression/verification/strictarithmetic/invalid/StrictArithmetic4.scala
@@ -1,0 +1,10 @@
+import leon.lang._
+
+object StrictArithmetic4 {
+
+  def foo4(x: Int, y: Int) = {
+    x % y
+  }
+
+}
+
